tests/wpt/css-tests/css-exclusions-1_dev/xhtml1print/wrap-flow-006.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
        <title>CSS Exclusions Test: wrap-flow property set to maximum</title>
        <link href="mailto:jacobg@adobe.com" rel="author" title="Jacob Goldstein" />
        <link href="http://www.w3.org/TR/css3-exclusions/#wrap-flow-property" rel="help" />
        <meta content="ahem dom" name="flags" />
        <meta content="Text should flow in the left OR right side of the exclusion depending on which has more space, as well as above and below" name="assert" />
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="resources/helper.js"></script>
        <style>
            * {
                margin: 0px;
                padding: 0px;
            }
            .exclusion {
                wrap-flow: maximum;
                position: absolute;
                top: 40px;
                left: 40px;
                width: 60px;
                height: 25px;
                background: blue;
            }
            #content {
                width: 400px;
                line-height: 12px;
                font: 12px Ahem;
            }
        </style>
    </head>
    <body>
        <div class="exclusion"></div>

        <div id="content">
            <span id="line1">ethical<br />beer<br />craft<br /></span>
            <span id="line2">foo</span>
            <span id="line3">sed</span>
            <span id="line4">Vegan<br />cliche<br />retro</span>
        </div>

        <div id="log"></div>

        <script type="text/javascript">
            /*Line 1*/
            test(function() {assert_equals(checkLinePos("line1",0,"top"), true)}, "Verify top of the 'line1' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line1",0,"left"), true)}, "Verify left of the 'line1' span is positioned correctly");
            /*Line 2*/
            test(function() {assert_equals(checkLinePos("line2",36,"top"), true)}, "Verify top of the 'line2' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line2",100,"left"), true)}, "Verify left of the 'line2' span is positioned correctly");
            /*Line 3*/
            test(function() {assert_equals(checkLinePos("line3",48,"top"), true)}, "Verify top of the 'line3' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line3",100,"left"), true)}, "Verify left of the 'line3' span is positioned correctly");
            /*Line 4*/
            test(function() {assert_equals(checkLinePos("line4",65,"top"), true)}, "Verify top of the 'line4' span is positioned correctly");
            test(function() {assert_equals(checkLinePos("line4",0,"left"), true)}, "Verify left of the 'line4' span is positioned correctly");
        </script>
    

</body></html>
